Copy current_data[i] to old_data[i] when it changes
Previously we just memcpy'ed the entrie current_data[]
to old_data[] whether it changed or not.  However, this caused
a bug that the data could very slowly change bit by bit and accumulate
a large amount of change without ever registering as having changed. If
we only copy the data when it changes, then small changes that accumulate
to large changes will register.